move the main code to src/main/java/com/app/atmsimulation
