+++ src/main/java/com/app/atmsimulation/Account.java
@@ -1,4 +1,4 @@
-package com.app;
+package com.app.atmsimulation;
 
 public class Account {
