+++ src/main/java/com/app/atmsimulation/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.app;
+package com.app.atmsimulation;
 
 import java.io.IOException;
 import java.nio.file.Files;
