+++ src/main/java/com/app/atmsimulation/TransactionScreen.java
@@ -1,4 +1,4 @@
-package com.app;
+package com.app.atmsimulation;
 
 import java.util.List;
 import java.util.Scanner;
