+++ src/main/java/com/app/atmsimulation/WelcomeScreen.java
@@ -1,4 +1,4 @@
-package com.app;
+package com.app.atmsimulation;
 
 import java.util.Scanner;
